requirements.txt
# Telegram
python-telegram-bot==20.7

# AI / LLM
anthropic==0.34.2
openai==1.40.0

# HTTP
requests==2.31.0
httpx==0.27.2
aiohttp==3.9.5
aiofiles==23.2.1

# Web scraping / research
beautifulsoup4==4.12.3
tavily-python==0.3.9
lxml==5.2.2

# Video editing
moviepy==1.0.3
ffmpeg-python==0.2.0
Pillow==10.4.0

# Google APIs
google-api-python-client==2.140.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1

# Utilities
python-dotenv==1.0.1
pydantic==2.8.2
rich==13.7.1
tenacity==8.3.0
